+++ starters/hello-world/src/pages/index.js
@@ -1,3 +1,5 @@
 import React from "react"
 
-export default () => <div>Hello world!</div>
+export default function Home() {
+  return <div>Hello world!</div>
+}